Create profoil-gfortran.bat
create bat file for windows user to create executable file and test the application. Need some feedback thought because the .tra file and the  .dump file are missing, but overall it works just like the .sh file. possible cause for the missing .tra file and the .dump file is the tee command. since windows doesn't have native tee alternative in cmd, i only dump the output of the executables to .txt file.